Bug fixed: New version of document now keeps files on HD from old version.

+++ dokeos/repository/lib/learning_object/document/document_form.class.php
@@ -106,7 +106,7 @@ class DocumentForm extends LearningObjectForm
 		}
 		elseif (strlen($_FILES['file']['name']) > 0)
 		{
-			unlink($this->get_upload_path().'/'.$object->get_path());
+			if ($values['version'] == 0) unlink($this->get_upload_path().'/'.$object->get_path());
 			$filename = $this->create_unique_filename($owner, $_FILES['file']['name']);
 			$path = $owner.'/'.$filename;
 			$full_path = $this->get_upload_path().'/'.$path;
